Encapsulate the Bayes searcher in another module
